requirements.txt
pyqt6
pyqt-tools
pyyaml
sqlalchemy
sqlalchemy.orm

